web-app/requirements.txt
flask==3.0.2
pymongo==4.6.2
python-dotenv==1.0.1
pytest==8.0.2
pytest-flask==1.3.0
pytest-cov==4.1.0
black==24.2.0
pylint==3.0.3
opencv-python==4.8.1.78
numpy==1.24.3
